Added /project/<pk>/refresh/ POST action.
Added a test case for the refresh action, which required refactoring the
testing and AWS mate instantiation procedures a little. The test case
does some monkeypatching but oh well.

Moved most of the data in initial_data to testing.yaml so the values
in the old initial_data would not mess up with test fixtures.