Moved tests -> systemtests to prevent 'manage.py test' from running them.
Refactored settings, freezr.settings.development is the default
settings module, but freezr.settings.base contains those that are
relevant to all environments. The mode is that a deployment package
would do "from freezr.settings.base import *" and fill in the
rest to make a full-blown application.

Added fixtures generated (and sanitized) from real CFN deployment.

Reorganized a lot of stuff. Freezr API, app, ui, core and backend are
listed now as separate apps with the idea that the UI might not be
running API, nor backend (celery) and so on.

"app" is really the "development app" and not a real production one.

Updated tests to match reorganization in the app structure.

./manage.py test -- pass
cd systemtests && ./run-integration-test.sh -- pass