Added pre-push script to check that things pass as they should in travis (see https://travis-ci.org/santtu/freezr/builds/15519548).
Also wondered for a long time why the pre-push script failed not
so deterministically. Finally figured out this was probably due to
a slow start of rabbitmq server due to log file growth.

'rabbitmqctl reset' took care of that, but also added wait loop
to freeze-thaw-aws.sh to wait until rabbitmq is answering before
proceeding (the failure was because POST to project tried to
issue account refresh, which could not be pushed to not-yet-responding
rabbitmq).